2022.3
The ugly HTML patching code had to be fixed.
Looks like now the prefix ends with <pre>.
Hope this doesn't break in some case.

Note that this version doesn't exactly match the
one uploaded to the Marketplace. The reason is
that 2022.2 was missing in the local repo clone,
so I mistakenly listed it in the changes as missing.
The code should be the same, though.